[Sonic Unleashed HUD] General refactoring and window title support
