--- DLLModsSource/SonicGenerations/SonicUnleashedHUD/UnleashedHUD/DataHelper.h
@@ -1,10 +1,0 @@
-#pragma once
-
-/**
-	Changes XNCP references to ones that use custom buttons and fix some bugs.
-**/
-void WriteData(int buttonType);
-
-void DoOnFrame();
-
-void HookFunctions();
