+++ DLLModsSource/SonicGenerations/SonicUnleashedHUD/UnleashedHUD/DllMain.h
@@ -1,0 +1,7 @@
+#pragma once
+
+class DllMain
+{
+public:
+	static HINSTANCE handle;
+};
