+++ DLLModsSource/SonicGenerations/SonicUnleashedHUD/UnleashedHUD/HudSonicStage.h
@@ -1,0 +1,8 @@
+#pragma once
+
+class HudSonicStage
+{
+public:
+	static void Install();
+};
+
