+++ DLLModsSource/SonicGenerations/SonicUnleashedHUD/UnleashedHUD/Mod.cpp
@@ -1,0 +1,8 @@
+extern "C" __declspec(dllexport) void Init(const char* path)
+{
+	Configuration::Read();
+
+	ControllerPatcher::Install();
+	HudSonicStage::Install();
+	Patches::Install();
+}
