+++ DLLModsSource/SonicGenerations/SonicUnleashedHUD/UnleashedHUD/Mod.ini
@@ -1,0 +1,18 @@
+[Desc]
+Author="PTKickass"
+Title="Sonic Unleashed HUD"
+Version="2.0"
+AuthorURL="https://www.youtube.com/user/ENPTKickass"
+Date="2020/02/20"
+Description="A Sonic Generations mod that partially ports the Sonic Unleashed HUD"
+
+[Main]
+UpdateServer="https://raw.githubusercontent.com/PTKickass/ModUpdates/master/SonicGenerations/SonicUnleashedHUD/"
+SaveFile=""
+ID="ptkickass.sonicgenerations.unleashedhud"
+IncludeDir0="."
+IncludeDirCount=1
+DependsCount=0
+DLLFile="UnleashedHUD.dll"
+CodeFile="UnleashedHUD.hmm"
+ConfigSchemaFile="UnleashedHUD.json"
