+++ DLLModsSource/SonicGenerations/SonicUnleashedHUD/UnleashedHUD/Patches.h
@@ -1,0 +1,7 @@
+#pragma once
+
+class Patches
+{
+public:
+	static void Install();
+};
