+++ DLLModsSource/SonicGenerations/SonicUnleashedHUD/UnleashedHUD/ScoreGenerations.ini
@@ -1,0 +1,2 @@
+[Developer]
+customXNCP=true
