+++ DLLModsSource/SonicGenerations/SonicUnleashedHUD/UnleashedHUD/UnleashedHUD.ini
@@ -1,0 +1,4 @@
+[Appearance]
+windowTitle=True
+windowTitleRegion=0
+buttonType=3
