+++ DLLModsSource/SonicGenerations/SonicUnleashedHUD/UnleashedHUD/UnleashedHUD.vcxproj
@@ -68,6 +68,13 @@
       <EnableUAC>false</EnableUAC>
       <AdditionalLibraryDirectories>$(ProjectDir)include\BlueBlur;%(AdditionalLibraryDirectories)</AdditionalLibraryDirectories>
     </Link>
+    <PostBuildEvent>
+      <Command>copy /Y "$(MSBuildProjectDirectory)\Mod.ini" "$(SolutionDir)$(Configuration)\mod.ini"
+copy /Y "$(MSBuildProjectDirectory)\ScoreGenerations.ini" "$(SolutionDir)$(Configuration)\ScoreGenerations.ini"
+copy /Y "$(MSBuildProjectDirectory)\UnleashedHUD.hmm" "$(SolutionDir)$(Configuration)\UnleashedHUD.hmm"
+copy /Y "$(MSBuildProjectDirectory)\UnleashedHUD.ini" "$(SolutionDir)$(Configuration)\UnleashedHUD.ini"
+copy /Y "$(MSBuildProjectDirectory)\UnleashedHUD.json" "$(SolutionDir)$(Configuration)\UnleashedHUD.json"</Command>
+    </PostBuildEvent>
   </ItemDefinitionGroup>
   <ItemDefinitionGroup Condition="'$(Configuration)|$(Platform)'=='Release|Win32'">
     <ClCompile>
@@ -97,27 +104,37 @@
       <EnableUAC>false</EnableUAC>
       <AdditionalLibraryDirectories>$(ProjectDir)include\BlueBlur;%(AdditionalLibraryDirectories)</AdditionalLibraryDirectories>
     </Link>
+    <PostBuildEvent>
+      <Command>copy /Y "$(MSBuildProjectDirectory)\Mod.ini" "$(SolutionDir)$(Configuration)\mod.ini"
+copy /Y "$(MSBuildProjectDirectory)\ScoreGenerations.ini" "$(SolutionDir)$(Configuration)\ScoreGenerations.ini"
+copy /Y "$(MSBuildProjectDirectory)\UnleashedHUD.hmm" "$(SolutionDir)$(Configuration)\UnleashedHUD.hmm"
+copy /Y "$(MSBuildProjectDirectory)\UnleashedHUD.ini" "$(SolutionDir)$(Configuration)\UnleashedHUD.ini"
+copy /Y "$(MSBuildProjectDirectory)\UnleashedHUD.json" "$(SolutionDir)$(Configuration)\UnleashedHUD.json"</Command>
+    </PostBuildEvent>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClInclude Include="DataHelper.h" />
+    <ClInclude Include="Configuration.h" />
+    <ClInclude Include="ControllerPatcher.h" />
     <ClInclude Include="detours\include\detours.h" />
     <ClInclude Include="detours\include\detver.h" />
     <ClInclude Include="detours\include\syelog.h" />
+    <ClInclude Include="DllMain.h" />
+    <ClInclude Include="HudSonicStage.h" />
     <ClInclude Include="include\Common.h" />
     <ClInclude Include="include\Helpers.h" />
     <ClInclude Include="include\INIReader.h" />
     <ClInclude Include="include\ModLoader.h" />
+    <ClInclude Include="Patches.h" />
     <ClInclude Include="Pch.h" />
+    <ClInclude Include="Resources.h" />
   </ItemGroup>
   <ItemGroup>
-    <ClCompile Include="controllers\Controller.cpp" />
-    <ClCompile Include="controllers\Controller.h" />
-    <ClCompile Include="controllers\Dualshock3.cpp" />
-    <ClCompile Include="controllers\Switch.cpp" />
-    <ClCompile Include="controllers\Xbox360.cpp" />
-    <ClCompile Include="controllers\XboxSeries.cpp" />
-    <ClCompile Include="DataHelper.cpp" />
-    <ClCompile Include="dllmain.cpp" />
+    <ClCompile Include="Configuration.cpp" />
+    <ClCompile Include="ControllerPatcher.cpp" />
+    <ClCompile Include="DllMain.cpp" />
+    <ClCompile Include="HudSonicStage.cpp" />
+    <ClCompile Include="Mod.cpp" />
+    <ClCompile Include="Patches.cpp" />
     <ClCompile Include="Pch.cpp">
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">Create</PrecompiledHeader>
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Release|Win32'">Create</PrecompiledHeader>
@@ -128,6 +145,12 @@
     <Library Include="detours\lib\x64\syelog.lib" />
     <Library Include="detours\lib\x86\detours.lib" />
     <Library Include="detours\lib\x86\syelog.lib" />
+  </ItemGroup>
+  <ItemGroup>
+    <ResourceCompile Include="UnleashedHUD.rc" />
+  </ItemGroup>
+  <ItemGroup>
+    <Image Include="Icon.ico" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
